add docstrings and breakout to own file
