+++ rubrik_cdm/cluster.py
@@ -1146,114 +1146,8 @@ class Cluster(Api):
             dict -- The full API response for `GET /v1/vmware/vcenter`.
         """
 
+        self.function_name = inspect.currentframe().f_code.co_name
+
         self.log(
             'get_all_vcenters: Getting information for each vCenter connected  to the Rubrik cluster.')
         return self.get('v1', '/vmware/vcenter', timeout=timeout)
-
-    def add_organization_protectable_object_sql_server_host(self, organization_name, mssql_host, timeout=15):
-
-        organization_id = self.object_id(
-            organization_name, "organization", timeout=timeout)
-
-        org_admin_id = self.object_id(
-            organization_name, "organization_admin_role", timeout=timeout)
-
-        self.log("add_organization_protectable_object_sql_server: Gathering the current MSSQL objects protected by the {} organization.".format(
-            organization_name))
-        current_mssql_protected_objects = self.get(
-            "internal", "/organization/{}/mssql".format(organization_id), timeout=timeout)
-
-        objects_to_protect = []
-
-        sql_host_id = self.object_id(
-            mssql_host, "physical_host", timeout=timeout)
-
-        for protected_object in current_mssql_protected_objects["data"]:
-            if protected_object["managedId"] == sql_host_id:
-                return "No change required. The MSSQL host {} is already assigned to the {} organization.".format(mssql_host, organization_name)
-
-        objects_to_protect.append(sql_host_id)
-
-        config = {
-            "authorizationSpecifications": [
-                {
-                    "privilege": "ManageRestoreSource",
-                    "resources": objects_to_protect
-                }
-            ],
-            "roleTemplate": "Organization"
-        }
-
-        return self.post("internal", "/role/{}/authorization".format(org_admin_id), config, timeout=timeout)
-
-    def add_organization_protectable_object_sql_server_db(self, organization_name, mssql_db, mssql_host, mssql_instance, timeout=15):
-
-        organization_id = self.object_id(
-            organization_name, "organization", timeout=timeout)
-
-        org_admin_id = self.object_id(
-            organization_name, "organization_admin_role", timeout=timeout)
-
-        self.log("add_organization_protectable_object_sql_server: Gathering the current MSSQL objects protected by the {} organization.".format(
-            organization_name))
-        current_mssql_protected_objects = self.get(
-            "internal", "/organization/{}/mssql".format(organization_id), timeout=timeout)
-
-        objects_to_protect = []
-
-        db_id = self.object_id(
-            mssql_db, "mssql_db", mssql_instance=mssql_instance, mssql_host=mssql_host, timeout=timeout)
-
-        for protected_object in current_mssql_protected_objects["data"]:
-            if protected_object["managedId"] == db_id:
-                return "No change required. The MSSQL DB {} is already assigned to the {} organization.".format(mssql_db, organization_name)
-
-        objects_to_protect.append(db_id)
-
-        config = {
-            "authorizationSpecifications": [
-                {
-                    "privilege": "ManageRestoreSource",
-                    "resources": objects_to_protect
-                }
-            ],
-            "roleTemplate": "Organization"
-        }
-
-        return self.post("internal", "/role/{}/authorization".format(org_admin_id), config, timeout=timeout)
-
-    def add_organization_protectable_object_sql_server_availability_group(self, organization_name, mssql_availability_group, timeout=15):
-
-        organization_id = self.object_id(
-            organization_name, "organization", timeout=timeout)
-
-        org_admin_id = self.object_id(
-            organization_name, "organization_admin_role", timeout=timeout)
-
-        self.log("add_organization_protectable_object_sql_server: Gathering the current MSSQL objects protected by the {} organization.".format(
-            organization_name))
-        current_mssql_protected_objects = self.get(
-            "internal", "/organization/{}/mssql".format(organization_id), timeout=timeout)
-
-        objects_to_protect = []
-
-        ag_id = self.object_id(
-            mssql_availability_group, "mssql_availability_group", timeout=timeout)
-
-        for protected_object in current_mssql_protected_objects["data"]:
-            if protected_object["managedId"] == ag_id:
-                return "No change required. The MSSQL Availability Group {} is already assigned to the {} organization.".format(mssql_availability_group, organization_name)
-
-        objects_to_protect.append(ag_id)
-
-        config = {
-            "authorizationSpecifications": [
-                {
-                    "privilege": "ManageRestoreSource",
-                    "resources": objects_to_protect
-                }
-            ],
-            "roleTemplate": "Organization"
-        }
-
-        return self.post("internal", "/role/{}/authorization".format(org_admin_id), config, timeout=timeout)
