+++ rubrik_cdm/rubrik_cdm.py
@@ -37,10 +37,11 @@ from .cluster import Cluster
 from .data_management import Data_Management
 from .physical import Physical
 from .cloud import Cloud
+from .organization import Organization
 from .exceptions import InvalidParameterException, RubrikException, APICallException, InvalidTypeException
 
 
-class Connect(Cluster, Data_Management, Physical, Cloud):
+class Connect(Cluster, Data_Management, Physical, Cloud, Organization):
     """This class acts as the base class for the Rubrik SDK and serves as the main interaction point
     for its end users. It also contains various helper functions used throughout the SDK.
